context - wip - refactoring
