+++ context.md
@@ -198,6 +198,65 @@ To manage this we run `Fetch` in a goroutine and it will write the result into a
 
 ## Refactor
 
+We can refactor our test code a bit by making assertion methods on our spy
+
+```go
+func (s *SpyStore) assertWasCancelled() {
+	s.t.Helper()
+	if !s.cancelled {
+		s.t.Errorf("store was not told to cancel")
+	}
+}
+
+func (s *SpyStore) assertWasNotCancelled() {
+	s.t.Helper()
+	if s.cancelled {
+		s.t.Errorf("store was told to cancel")
+	}
+}
+```
+
+Remember to pass in the `*testing.T` when creating the spy. 
+
+```go
+func TestServer(t *testing.T) {
+	data := "hello, world"
+
+	t.Run("returns data from store", func(t *testing.T) {
+		store := &SpyStore{response: data, t: t}
+		svr := Server(store)
+
+		request := httptest.NewRequest(http.MethodGet, "/", nil)
+		response := httptest.NewRecorder()
+
+		svr.ServeHTTP(response, request)
+
+		if response.Body.String() != data {
+			t.Errorf(`got "%s", want "%s"`, response.Body.String(), data)
+		}
+
+		store.assertWasNotCancelled()
+	})
+
+	t.Run("tells store to cancel work if request is cancelled", func(t *testing.T) {
+		store := &SpyStore{response: data, t: t}
+		svr := Server(store)
+
+		request := httptest.NewRequest(http.MethodGet, "/", nil)
+
+		cancellingCtx, cancel := context.WithCancel(request.Context())
+		time.AfterFunc(5*time.Millisecond, cancel)
+		request = request.WithContext(cancellingCtx)
+
+		response := httptest.NewRecorder()
+
+		svr.ServeHTTP(response, request)
+
+		store.assertWasCancelled()
+	})
+}
+```
+
 
 ---------
