+++ context/v2/context.go
@@ -1,0 +1,30 @@
+package v1
+
+import (
+	"fmt"
+	"net/http"
+)
+
+type Store interface {
+	Fetch() string
+	Cancel()
+}
+
+func Server(store Store) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		ctx := r.Context()
+
+		data := make(chan string, 1)
+
+		go func() {
+			data <- store.Fetch()
+		}()
+
+		select {
+		case d := <-data:
+			fmt.Fprint(w, d)
+		case <-ctx.Done():
+			store.Cancel()
+		}
+	}
+}
