+++ context/v2/testdoubles.go
@@ -1,0 +1,35 @@
+package v1
+
+import (
+	"testing"
+	"time"
+)
+
+type SpyStore struct {
+	response  string
+	cancelled bool
+	t *testing.T
+}
+
+func (s *SpyStore) Fetch() string {
+	time.Sleep(100 * time.Millisecond)
+	return s.response
+}
+
+func (s *SpyStore) Cancel() {
+	s.cancelled = true
+}
+
+func (s *SpyStore) assertWasCancelled() {
+	s.t.Helper()
+	if !s.cancelled {
+		s.t.Errorf("store was not told to cancel")
+	}
+}
+
+func (s *SpyStore) assertWasNotCancelled() {
+	s.t.Helper()
+	if s.cancelled {
+		s.t.Errorf("store was told to cancel")
+	}
+}
